MOC-63: [BE] API for change groupmember type

+++ mocker-back/src/main/java/com/mocker/service/impl/GroupMemberServiceImpl.java
@@ -1,6 +1,8 @@
 package com.mocker.service.impl;
 
+import com.mocker.configuration.security.ApplicationContextHolder;
 import com.mocker.domain.exception.NotFoundException;
+import com.mocker.domain.exception.PermissionException;
 import com.mocker.domain.model.entity.Group;
 import com.mocker.domain.model.entity.GroupMember;
 import com.mocker.domain.model.entity.User;
@@ -13,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
 import java.util.UUID;
 
 /**
@@ -25,14 +28,22 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class GroupMemberServiceImpl implements GroupMemberService {
 
+    private final ApplicationContextHolder applicationContextHolder;
     private final GroupMemberRepository groupMemberRepository;
     private final GroupRepository groupRepository;
     private final UserRepository userRepository;
 
     @Override
     public GroupMember delete(GroupMember groupMember) {
+        UUID authId = applicationContextHolder.getCurrentUser().getId();
         if (groupMemberRepository.findById(groupMember.getId()).isEmpty()) {
-            throw new NotFoundException("GroupMember with id " + groupMember.getId() + " validation.data_not_found");
+            throw new NotFoundException(groupMember.getId().getGroupId());
+        }
+        if ((!groupRepository.getRoleUserInGroup(groupMember.getGroup().getId(), authId).equals("GROUP_ADMIN")
+                || !groupRepository.getRoleUserInGroup(groupMember.getGroup().getId(), authId).equals("GROUP_ASSOCIATE"))
+                || (groupRepository.getRoleUserInGroup(groupMember.getGroup().getId(), groupMember.getUser().getId()).equals("GROUP_ADMIN")
+                && groupRepository.getRoleUserInGroup(groupMember.getGroup().getId(), authId).equals("GROUP_ADMIN"))) {
+            throw new PermissionException(authId);
         }
         groupMemberRepository.deleteById(groupMember.getId());
         return groupMember;
@@ -40,16 +51,38 @@ public class GroupMemberServiceImpl implements GroupMemberService {
 
     @Override
     public GroupMember upsert(GroupMember groupMember) {
+        UUID authId = applicationContextHolder.getCurrentUser().getId();
+        boolean isNew = groupMember.getId() == null;
         UUID groupId = groupMember.getGroup().getId();
-        Group group = groupRepository.findById(groupId).orElseThrow(() -> new NotFoundException("Group with id: " + groupId + " validation.data_not_found"));
+        Group group = groupRepository.findById(groupId).orElseThrow(() -> new NotFoundException(groupId));
         UUID userId = groupMember.getUser().getId();
-        User user = userRepository.findById(userId).orElseThrow(() -> new NotFoundException("User with id: " + userId + " validation.data_not_found"));
-        groupMemberRepository.save(GroupMember.builder()
-                .id(groupMember.getId())
-                .group(group)
-                .user(user)
-                .role(Role.GROUP_ADMIN)
-                .build());
+        User user = userRepository.findById(userId).orElseThrow(() -> new NotFoundException(userId));
+        if (!groupRepository.getRoleUserInGroup(groupId, authId).equals("GROUP_ADMIN")
+                || !groupRepository.getRoleUserInGroup(group.getId(), authId).equals("GROUP_ASSOCIATE")) {
+            throw new PermissionException(authId);
+        }
+        if (isNew) {
+            groupMemberRepository.save(GroupMember.builder()
+                    .id(groupMember.getId())
+                    .group(group)
+                    .user(user)
+                    .role(Role.USER)
+                    .build());
+        } else {
+            String oldRole = groupRepository.getRoleUserInGroup(groupId, userId);
+            String newRole = groupMember.getRole().toString();
+            //only role Group_admin can change role
+            if (!Objects.equals(oldRole, newRole) && !groupRepository.getRoleUserInGroup(groupId, authId).equals("GROUP_ADMIN")) {
+                throw new PermissionException(authId);
+            }
+            groupMemberRepository.save(GroupMember.builder()
+                    .id(groupMember.getId())
+                    .group(group)
+                    .user(user)
+                    .role(Role.valueOf(newRole))
+                    .build());
+        }
         return groupMember;
     }
+
 }
